Moved offset into NurbsCurve. Removed Offset class and OffsetTests.
